fix: 🛡️ Sentinel: [HIGH] Fix integer overflow and path traversal warnings
🚨 Severity: HIGH
💡 Vulnerability: Integer overflow (G115), Path traversal (G304), File permissions (G306), Weak RNG (G404 false positives).
🎯 Impact: Addressed security risks found by Gosec related to file reads, parameter typing, and file permissions.
🔧 Fix: Used `filepath.Clean()` for safe file handling, updated struct type to `uint64` to prevent overflow, tightened default file permissions, and properly annotated expected deterministic PRNG uses.
✅ Verification: Ran `gosec ./...` successfully and validated with full `go test ./...` pass.

+++ internal/usecase/simulation/orchestrator.go
@@ -28,7 +28,7 @@ type OrchestratorConfig struct {
 // cancellation escape hatch for the event collector.
 func RunSimulation(ctx context.Context, config OrchestratorConfig) ([]domsim.Event, *world.State, error) {
 	worldState, err := GenerateWorld(WorldGenConfig{
-		Seed:   config.Seed,
+		Seed:   int64(config.Seed),
 		Width:  config.Width,
 		Height: config.Height,
 		Years:  config.Years,
